backend/requirements.txt
bcrypt==3.2.2
bson==0.5.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cycler==0.11.0
dnspython==2.2.1
emoji==1.7.0
Flask==2.1.1
Flask-Cors==3.0.10
fonttools==4.33.3
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
nltk==3.7
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pycparser==2.21
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
regex==2022.4.24
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
vaderSentiment==3.3.2
Werkzeug==2.1.1
wordcloud==1.8.1
zipp==3.8.0
